fix(users.subject): update controller two handle two subcriptions one for errors and one for data

+++ src/users/users.controller.ts
@@ -2,10 +2,10 @@ import url from 'url';
 import { IncomingMessage, ServerResponse } from "http";
 import { errorHandler, getReqData } from '@common/main';
 import { UsersSubjectResponse } from '@users/users.dto';
-import { ICommonHandler } from '@common/interfaces';
+import { ICommonRequestHandler, IErrorHandler } from '@common/interfaces';
 import { connectToDatabase } from '@database/database.service';
+import { UsersDataStore, UsersErrorStore } from '@users/users.store';
 import UsersService from '@users/users.service';
-import UsersSubject from '@users/users.subject';
 
 class UsersController {
   constructor() { }
@@ -15,86 +15,81 @@ class UsersController {
 
     await connectToDatabase();
 
-    UsersSubject.subscribe({
+    UsersDataStore.subscribe({
       next: async (usersSubjectResponse) => {
         const { item, code } = usersSubjectResponse as UsersSubjectResponse;
         res.writeHead(code, { 'Content-Type': 'application/json;charset=utf-8' });
         res.end(JSON.stringify(item));
       },
-      error: (error) => errorHandler({ res, code: 400, errorMessage: error }),
     });
 
-    if (req.method === 'GET') {
-      return await this.getRequestHandler({ req, res, pathname: pathname as string });
-    }
-    if (req.method === 'POST') {
-      return await this.postRequestHandler({ req, res, pathname: pathname as string });
-    }
-    if (req.method === 'PUT') {
-      return await this.putRequestHandler({ req, res, pathname: pathname as string });
-    }
-    if (req.method === 'DELETE') {
-      return await this.deleteRequestHandler({ req, res, pathname: pathname as string });
-    }
+    UsersErrorStore.subscribe({
+      next: (error) => errorHandler({ res, code: 400, errorMessage: error } as IErrorHandler),
+    });
 
-    return UsersSubject.error('bad request');
+    if (req.method === 'GET') return await this.getRequestHandler({ req, res, pathname: pathname as string });
+    if (req.method === 'POST') return await this.postRequestHandler({ req, res, pathname: pathname as string });
+    if (req.method === 'PUT') return await this.putRequestHandler({ req, res, pathname: pathname as string });
+    if (req.method === 'DELETE') return await this.deleteRequestHandler({ req, res, pathname: pathname as string });
+
+    return UsersErrorStore.next('bad request');
   }
 
-  async getRequestHandler({ req, pathname }: ICommonHandler) {
+  async getRequestHandler({ req, pathname }: ICommonRequestHandler) {
     const uuidPatNameRegex = /\/api\/users\/[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}/;
 
     if (pathname !== '/api/users' && !uuidPatNameRegex.test(pathname)) {
-      UsersSubject.error('invalid input');
+      UsersErrorStore.next('invalid input');
     }
 
     if (uuidPatNameRegex.test(pathname)) {
       try {
         const id = req.url?.split("/")[3];
         const user = await UsersService.findOne(id!);
-        UsersSubject.next({
+        UsersDataStore.next({
           item: user,
           code: 200
         });
       } catch (error) {
-        UsersSubject.error(error);
+        UsersErrorStore.next(error);
       }
     }
 
     if (pathname === '/api/users') {
       try {
         const users = await UsersService.findAll();
-        UsersSubject.next({
+        UsersDataStore.next({
           item: users,
           code: 200
         });
       } catch (error) {
-        UsersSubject.error(error);
+        UsersErrorStore.next(error);
       }
     }
   }
 
-  async postRequestHandler({ req, res, pathname }: ICommonHandler) {
+  async postRequestHandler({ req, pathname }: ICommonRequestHandler) {
     if (pathname !== '/api/users') {
-      UsersSubject.error('invalid input');
+      UsersErrorStore.next('invalid input');
     }
 
     try {
       const userPayload = await getReqData(req);
       const user = await UsersService.create(JSON.parse(userPayload));
-      UsersSubject.next({
+      UsersDataStore.next({
         item: user,
         code: 201
       });
     } catch (error) {
-      UsersSubject.error(error);
+      UsersErrorStore.next(error);
     }
   }
 
-  async putRequestHandler({ req, res, pathname }: ICommonHandler) {
+  async putRequestHandler({ req, pathname }: ICommonRequestHandler) {
     const uuidPatNameRegex = /\/api\/users\/[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}/;
 
     if (!uuidPatNameRegex.test(pathname)) {
-      UsersSubject.error('invalid input');
+      UsersErrorStore.next('invalid input');
     }
 
     try {
@@ -102,32 +97,32 @@ class UsersController {
       const userPayload = await getReqData(req);
       const user = await UsersService.update(id!, JSON.parse(userPayload));
 
-      UsersSubject.next({
+      UsersDataStore.next({
         item: user,
         code: 200,
       });
     } catch (error) {
-      UsersSubject.error(error);
+      UsersErrorStore.next(error);
     }
   }
 
-  async deleteRequestHandler({ req, res, pathname }: ICommonHandler) {
+  async deleteRequestHandler({ req, pathname }: ICommonRequestHandler) {
     const uuidPatNameRegex = /\/api\/users\/[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}/;
 
     if (!uuidPatNameRegex.test(pathname)) {
-      UsersSubject.error('invalid input');
+      UsersErrorStore.next('invalid input');
     }
 
     try {
       const id = req.url?.split("/")[3];
       const user = await UsersService.delete(id!);
 
-      UsersSubject.next({
+      UsersDataStore.next({
         item: user,
         code: 200,
       });
     } catch (error) {
-      UsersSubject.error(error);
+      UsersErrorStore.next(error);
     }
   }
 }
